requirements.txt
numpy~=1.22.2
torch~=1.10.2
albumentations~=1.1.0
Pillow~=9.0.1
torchvision~=0.11.3
torchmetrics~=0.7.2
matplotlib~=3.5.1
seaborn~=0.11.2
scikit-learn
wandb
scipy
tqdm